Remove worker/scripts/normalize-compile-commands.mjs file and integrate it in worker/scripts/clang-scripts.mjs in a new npm run normalize-compile-commands NPM script. Why? Because it doesn't depend on cwd. It generates the same correct normalized compile_commands.json file no matter where npm run normalize-compile-commands is called from
