+++ worker/scripts/clang-scripts.mjs
@@ -1,5 +1,7 @@
-import * as path from 'node:path';
+import * as process from 'node:process';
 import * as os from 'node:os';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import { execSync } from 'node:child_process';
 import { globSync } from 'glob';
 
@@ -50,6 +52,12 @@ async function run() {
 
 		case 'tidy:fix': {
 			tidy({ fix: true });
+
+			break;
+		}
+
+		case 'normalize-compile-commands': {
+			normalizeCompileCommands();
 
 			break;
 		}
@@ -139,6 +147,35 @@ function tidy({ fix }) {
 	executeCmd(
 		`"${PYTHON}" "${runClangTidy}" -clang-tidy-binary="${clangTidy}" -clang-apply-replacements-binary="${clangApplyReplacements}" -p="${BUILD_DIR}" -j=${NUM_CORES} -quiet ${fixArg} -format ${tidyChecksArg} ${runClangTidyFilesArgs}`
 	);
+}
+
+function normalizeCompileCommands() {
+	logInfo('normalizeCompileCommands()');
+
+	const rootDir = path.resolve(path.join('../../'));
+	const compileCommandsFile = `${BUILD_DIR}/compile_commands.json`;
+
+	try {
+		const commands = JSON.parse(fs.readFileSync(compileCommandsFile, 'utf8'));
+
+		for (const entry of commands) {
+			if (entry.file && entry.directory) {
+				// Resolve to absolute path first.
+				const absolutePath = path.resolve(entry.directory, entry.file);
+
+				// Convert to relative path from repo root.
+				entry.file = path.relative(rootDir, absolutePath);
+			}
+		}
+
+		fs.writeFileSync(compileCommandsFile, JSON.stringify(commands, null, 2));
+	} catch (error) {
+		logError(
+			`normalizeCompileCommands() | failed to clean up compile_commands.json: ${error}`
+		);
+
+		exitWithError();
+	}
 }
 
 function getClangToolBinary({ clangToolName, version, checkRequireVersion }) {
